requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aioitertools==0.12.0
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.17.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
arize-phoenix==10.13.1
arize-phoenix-client==1.18.0
arize-phoenix-evals==0.29.0
arize-phoenix-otel==0.13.0
astroid==4.0.1
asttokens==3.0.1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.4.0
audio-recorder-streamlit==0.0.8
Authlib==1.6.5
backoff==2.2.1
backports.asyncio.runner==1.2.0
black==25.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
cryptography==45.0.7
datasets==4.4.1
debugpy==1.8.17
decorator==5.2.1
dependency-injector==4.48.2
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
Faker==38.2.0
fastapi==0.116.1
fasttext-wheel==0.9.2
fastuuid==0.14.0
filelock==3.20.0
flashtext==2.7
frozenlist==1.8.0
fsspec==2025.10.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
googleapis-common-protos==1.70.0
graphql-core==3.2.6
greenlet==3.2.4
griffe==1.14.0
grpc-interceptor==0.15.4
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hazm==0.10.0
hf-xet==1.2.0
holidays==0.85
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.4.0
iniconfig==2.3.0
ipykernel==7.1.0
ipython==8.37.0
ipywidgets==8.1.8
isort==7.0.0
jalali_core==1.0.0
jdatetime==5.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kubernetes==33.1.0
litellm==1.79.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mccabe==0.7.0
mcp==1.17.0
mdurl==0.1.2
mem0ai==1.0.0
monotonic==1.6
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==2.10.1
neo4j==5.28.2
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.2
nodeenv==1.9.1
numpy==1.24.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
openai==2.8.1
openai-agents==0.6.0
openinference-instrumentation==0.1.38
openinference-instrumentation-openai==0.1.31
openinference-instrumentation-openai-agents==1.2.0
openinference-semantic-conventions==0.1.21
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.48b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.3
ormsgpack==1.12.0
packaging==24.2
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.5.0
pluggy==1.6.0
ply==3.11
portalocker==3.2.0
posthog==3.25.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==5.29.5
psutil==7.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==3.0.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.10.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pylint==4.0.2
pyright==1.1.407
pystache==0.6.8
pytest==9.0.2
pytest-asyncio==1.3.0
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytokens==0.2.0
pytz==2024.2
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.15.1
redis==6.4.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.13.1
sentence-transformers==3.3.1
simhash==2.1.2
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlean.py==3.50.4.4
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.47.2
strawberry-graphql==0.270.1
streamlit==1.39.0
streamlit-autorefresh==1.0.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
tomlkit==0.13.3
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.5.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==5.0.3
wcwidth==0.2.14
websocket-client==1.9.0
widgetsnbextension==4.0.15
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
